cloud-networking/requirements.txt
google-api-python-client
google-cloud-storage
gcloud-aio-storage
boto3
tomli
pyyaml
openpyxl
guppy3
